changed stock lastDiv and Purchase to double

+++ FinPulse.BL/Dtos/Stock/StockCreateDto.cs
@@ -12,10 +12,10 @@ public class StockCreateDto
     public string CompanyName { get; set; } = string.Empty;
     [Required]
     [Range(1,1000000000)]
-    public decimal Purchase { get; set; }
+    public double Purchase { get; set; }
     [Required]
     [Range(0,100)]
-    public decimal LastDiv { get; set; }
+    public double LastDiv { get; set; }
     [Required]
     [MaxLength(10, ErrorMessage = "Industry cannot be over 10 characters.")]
     public string Industry { get; set; } = string.Empty;
